pyproject.toml
# ~~ Generated by projen. To modify, edit .projenrc.py and run "npx projen".

[tool.poetry]
name = "llama_index_in_lambda"
version = "0.1.0"
description = ""
authors = [ "ArJSarmiento <rneljan@gmail.com>" ]
readme = "README.md"

  [tool.poetry.dependencies]
  python = ">=3.11, <3.13"
  llama-index = "0.11.1"
  fastapi = {extras = ["standard"], version = "^0.112.2"}
  llama-index-embeddings-bedrock = "^0.3.0"
  llama-index-vector-stores-pinecone = "^0.2.1"
  llama-index-llms-bedrock = "^0.2.1"
  uvicorn = "^0.30.6"

[tool.poetry.group.format.dependencies]
ruff = "0.5.1"

[tool.poetry.group.dev.dependencies]
projen = "0.86.7"
pytest = "7.4.3"
boto3 = "^1.35.5"
aws-cdk-lib = "^2.1.0"
constructs = "^10.0.5"
aws-solutions-constructs-aws-apigateway-lambda = "^2.65.0"
aws-cdk-aws-lambda-python-alpha = "^2.156.0a0"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"
